+++ src/main/java/com/educandoweb/curso/entities/Order.java
@@ -19,7 +19,7 @@ public class Order implements Serializable{
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private Instant moment;
 
 	@ManyToOne
@@ -30,18 +30,18 @@ public class Order implements Serializable{
 		
 	}
 
-	public Order(Integer id, Instant moment, User client) {
+	public Order(Long id, Instant moment, User client) {
 		super();
 		this.id = id;
 		this.moment = moment;
 		this.client = client;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
